feat: CommonErrorCode 추가 및 디렉토리 이동
