+++ src/main/java/com/sparta/hhplusconcert/common/exception/TokenErrorCode.java
@@ -1,4 +1,4 @@
-package com.sparta.hhplusconcert.exception;
+package com.sparta.hhplusconcert.common.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -8,8 +8,8 @@ import org.springframework.http.HttpStatus;
 @RequiredArgsConstructor
 public enum TokenErrorCode implements  ErrorCode{
 
-  EXPIRED_WAITING_TOKEN(HttpStatus.UNAUTHORIZED, "Token has expired"),
-  INVALID_TOKEN(HttpStatus.UNAUTHORIZED, "Token is invalid")
+  EXPIRED_WAITING_TOKEN(HttpStatus.UNAUTHORIZED, "Waiting token has expired"),
+  INVALID_WAITING_TOKEN(HttpStatus.UNAUTHORIZED, "Waiting token is invalid")
   ;
 
   private final HttpStatus httpStatus;
