--- src/main/java/com/sparta/hhplusconcert/concert/domain/ExpiredTokenException.java
@@ -1,7 +1,0 @@
-package com.sparta.hhplusconcert.concert.domain;
-
-public class ExpiredTokenException extends RuntimeException {
-  public ExpiredTokenException(String message) {
-    super(message);
-  }
-}
