+++ src/main/java/com/sparta/hhplusconcert/concert/domain/InvalidTokenException.java
@@ -1,6 +1,6 @@
 package com.sparta.hhplusconcert.concert.domain;
 
-import com.sparta.hhplusconcert.exception.ErrorCode;
+import com.sparta.hhplusconcert.common.exception.ErrorCode;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
